+++ server/API/Infrastructure/Authentication/ApiKeyStore.cs
@@ -4,7 +4,7 @@ public static class ApiKeyStore
 {
     public static ApiKey[] Keys { get; private set; } = Array.Empty<ApiKey>();
 
-    public static void InitKeys(IEnumerable<string> keys, int requestLimit, int requestLimitPeriodHours)
+    public static void InitKeys(IEnumerable<string> keys, int requestLimit, double requestLimitPeriodHours)
     {
         Keys = keys.Select(k => new ApiKey(k, requestLimit, requestLimitPeriodHours)).ToArray();
     }
